Revert "delete CameraView"
This reverts commit 06a7fe7cdc1072e9c602507af013736d73eb61df.
